--- salt/utils/linux_osinfo.py
@@ -1,6 +1,0 @@
-def linux_distribution():
-    """ """
-    # Late import so that when getting called from setup.py does not break
-    import distro
-
-    return distro.id(), distro.version(best=True), distro.codename()
